@@ -1,6 +1,6 @@
 class Error {
   final int code;
-  final String message;
+  final String? message;
 
   /// Custom error code.
   Error.custom(this.code, [this.message]);
