@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:centrifuge/centrifuge.dart';
+import 'package:centrifuge/centrifuge.dart' hide Error;
 import 'package:centrifuge/src/client.dart';
-import 'package:centrifuge/src/proto/client.pb.dart' hide Error;
+import 'package:centrifuge/src/proto/client.pb.dart';
 import 'package:centrifuge/src/transport.dart';
 import 'package:mockito/mockito.dart';
 import 'package:protobuf/protobuf.dart';
@@ -14,10 +14,10 @@ class MockWebSocket implements WebSocket {
   final _stubs = <CommandMatcher, SendAction>{};
 
   @override
-  Duration pingInterval;
+  Duration? pingInterval;
 
   @override
-  String closeReason;
+  String? closeReason;
 
   @override
   void add(dynamic data) {
@@ -28,10 +28,10 @@ class MockWebSocket implements WebSocket {
     for (CommandMatcher func in _stubs.keys) {
       if (func(command)) {
         final reply = Reply()..id = command.id;
-        if (_stubs[func]._error != null) {
-          reply.error = _stubs[func]._error;
-        } else if (_stubs[func].result != null) {
-          reply.result = _stubs[func]._result.writeToBuffer();
+        if (_stubs[func]!._error != null) {
+          reply.error = _stubs[func]!._error as Error;
+        } else if (_stubs[func]!.result != null) {
+          reply.result = _stubs[func]!._result.writeToBuffer();
         }
 
         final replyData = reply.writeToBuffer();
@@ -39,7 +39,7 @@ class MockWebSocket implements WebSocket {
 
         final writer = CodedBufferWriter();
         writer.writeInt32NoTag(length);
-        onData(writer.toBuffer() + replyData);
+        onData!(writer.toBuffer() + replyData);
       }
     }
 
@@ -52,24 +52,33 @@ class MockWebSocket implements WebSocket {
     return action;
   }
 
-  Function onData;
-  Function onError;
-  Function onDone;
+  Function? onData;
+  Function? onError;
+  Function? onDone;
 
   @override
-  StreamSubscription listen(void onData(dynamic event),
-      {Function onError, void onDone(), bool cancelOnError}) {
+  StreamSubscription listen(void onData(dynamic event)?, {Function? onError, void onDone()?, bool? cancelOnError}) {
     this.onData = onData;
     this.onError = onError;
     this.onDone = onDone;
-    return null;
+
+    final subscription = StreamController<void>();
+
+    return subscription.stream.listen(
+      (event) => onDone,
+      onError: onError,
+      onDone: onDone,
+    );
   }
 
   @override
-  Future close([int code, String reason]) {
-    closeReason = reason;
-    onDone();
-    return null;
+  Future close([int? code, String? reason]) {
+    return Future<void>(
+      () {
+        closeReason = reason;
+        onDone!();
+      },
+    );
   }
 
   @override
@@ -77,8 +86,8 @@ class MockWebSocket implements WebSocket {
 }
 
 class SendAction {
-  GeneratedMessage _error;
-  GeneratedMessage _result;
+  GeneratedMessage? _error;
+  late GeneratedMessage _result;
 
   void result(GeneratedMessage result) {
     _result = result;
@@ -91,62 +100,56 @@ class SendAction {
 
 typedef CommandMatcher = Function(Command);
 
-CommandMatcher withMethod(MethodType type) =>
-    (Command command) => command.method == type;
+CommandMatcher withMethod(MethodType type) => (Command command) => command.method == type;
 
 class MockTransport implements Transport {
-  String url;
-  TransportConfig transportConfig;
+  String? url;
+  TransportConfig? transportConfig;
 
   @override
-  Future close() {
+  Future? close() {
     // TODO: implement close
     return null;
   }
 
-  void Function(Push push) onPush;
-  void Function(dynamic) onError;
-  void Function(String, bool) onDone;
+  void Function(Push push)? onPush;
+  void Function(dynamic)? onError;
+  void Function(String?, bool)? onDone;
 
-  Completer<void> _openCompleter;
+  Completer<void>? _openCompleter;
 
   void completeOpen() {
     assert(_openCompleter != null);
 
-    _openCompleter.complete();
+    _openCompleter!.complete();
     _openCompleter = null;
   }
 
   void completeOpenError(dynamic error) {
     assert(_openCompleter != null);
 
-    _openCompleter.completeError(error);
+    _openCompleter!.completeError(error);
     _openCompleter = null;
   }
 
   @override
-  Future open(void Function(Push push) onPush,
-      {Function onError, void Function(String, bool) onDone}) {
+  Future open(void Function(Push push) onPush, {Function? onError, void Function(String?, bool)? onDone}) {
     assert(_openCompleter == null);
     _openCompleter = Completer<void>.sync();
 
     this.onPush = onPush;
-    this.onError = onError;
+    this.onError = onError as void Function(dynamic)?;
     this.onDone = onDone;
 
-    return _openCompleter.future;
+    return _openCompleter!.future;
   }
 
   final sendList = <Triplet>[];
 
-  Triplet<Req, Res> sendListLast<Req extends GeneratedMessage,
-          Res extends GeneratedMessage>() =>
-      sendList.last;
+  Triplet<Req, Res> sendListLast<Req extends GeneratedMessage, Res extends GeneratedMessage>() => sendList.last as Triplet<Req, Res>;
 
   @override
-  Future<Rep>
-      sendMessage<Req extends GeneratedMessage, Rep extends GeneratedMessage>(
-          Req request, Rep result) {
+  Future<Rep> sendMessage<Req extends GeneratedMessage, Rep extends GeneratedMessage>(Req request, Rep result) {
     final completer = Completer<Rep>.sync();
 
     sendList.add(Triplet<Req, Rep>(request, result, completer));
@@ -164,8 +167,7 @@ class Triplet<Req extends GeneratedMessage, Res extends GeneratedMessage> {
 
   void completeWith(Res result) => completer.complete(result);
 
-  void completeWith2(Res Function(Res) updates) =>
-      completer.complete(updates(result));
+  void completeWith2(Res Function(Res) updates) => completer.complete(updates(result));
 
   void completeWithError(dynamic error) => completer.completeError(error);
 
@@ -177,9 +179,7 @@ class Triplet<Req extends GeneratedMessage, Res extends GeneratedMessage> {
   @override
   bool operator ==(dynamic other) {
     if (other is! Triplet) return false;
-    return other.request == request &&
-        other.completer == completer &&
-        other.completer == completer;
+    return other.request == request && other.completer == completer && other.completer == completer;
   }
 
   @override
@@ -188,8 +188,8 @@ class Triplet<Req extends GeneratedMessage, Res extends GeneratedMessage> {
 
 class MockClient extends Mock implements ClientImpl {
   @override
-  ClientConfig config;
+  ClientConfig? config;
 
   @override
-  bool connected;
+  bool? connected;
 }
